+++ src/util.ts
@@ -262,6 +262,33 @@ export async function safelyExecute(operation: () => Promise<any>, logError = fa
 }
 
 /**
+ * Execute and return an asynchronous operation with a timeout
+ *
+ * @param operation - Function returning a Promise
+ * @param logError - Determines if the error should be logged
+ * @param retry - Function called if an error is caught
+ * @param timeout - Timeout to fail the operation
+ * @returns - Promise resolving to the result of the async operation
+ */
+export async function safelyExecuteWithTimeout(operation: () => Promise<any>, logError = false, timeout = 500) {
+  try {
+    return await Promise.race([
+      operation(),
+      new Promise<void>((_, reject) =>
+        setTimeout(() => {
+          reject(new Error('timeout'));
+        }, timeout),
+      ),
+    ]);
+  } catch (error) {
+    /* istanbul ignore next */
+    if (logError) {
+      console.error(error);
+    }
+  }
+}
+
+/**
  * Validates a Transaction object for required properties and throws in
  * the event of any validation error.
  *
@@ -537,6 +564,7 @@ export async function estimateGas(transaction: Transaction, ethQuery: any) {
 export default {
   BNToHex,
   fractionBN,
+  query,
   getBuyURL,
   handleFetch,
   hexToBN,
@@ -545,6 +573,7 @@ export default {
   constructTxParams,
   normalizeTransaction,
   safelyExecute,
+  safelyExecuteWithTimeout,
   successfulFetch,
   timeoutFetch,
   validateTokenToWatch,
@@ -553,5 +582,4 @@ export default {
   validateTypedSignMessageDataV3,
   calcTokenAmount,
   estimateGas,
-  query,
 };